+++ BGAnimations/ScreenSelectStyle underlay/choice.lua
@@ -10,16 +10,9 @@ local af = Def.ActorFrame{
     InitCommand=function(self)
         self:aux(0)
         self:zoom(0.5):xy( frame_x, _screen.cy + WideScale(0,10) )
-
-        if ThemePrefs.Get("VisualStyle")=="Gay" and not HolidayCheer() then
-            self:bob():effectmagnitude(0,0,0):effectclock('bgm'):effectperiod(0.666)
-        end
     end,
     GainFocusCommand=function(self)
         self:finishtweening():linear(0.125):zoom(1)
-        if ThemePrefs.Get("VisualStyle")=="Gay" and not HolidayCheer() then
-            self:effectmagnitude(0,4,0)
-        end
     end,
     LoseFocusCommand=function(self)
         self:finishtweening():linear(0.125):zoom(0.5):effectmagnitude(0,0,0)
@@ -43,8 +36,6 @@ local af = Def.ActorFrame{
         InitCommand=function(self)
             self:shadowlength(1):y(37):zoom(0.5)
         end,
-        GainFocusCommand=function(self) if ThemePrefs.Get("VisualStyle")=="Gay" then self:rainbowscroll(true) end end,
-        LoseFocusCommand=function(self) if ThemePrefs.Get("VisualStyle")=="Gay" then self:rainbowscroll(false) end end,
     }
 }