+++ BGAnimations/_shared background/default.lua
@@ -3,8 +3,6 @@
 -- 	return LoadActor( THEME:GetPathB("", "_shared background/Snow.lua") )
 -- end
 
-local file = THEME:GetPathG("", "VOLT26/SharedBackground.png")
-
 local af = Def.ActorFrame{}
 
 -- a simple Quad to serve as the backdrop
@@ -12,9 +10,12 @@ af[#af+1] = Def.Quad{
 	InitCommand=function(self) self:FullScreen():Center():diffuse( ThemePrefs.Get("RainbowMode") and Color.White or Color.Black ) end,
 }
 
-af[#af+1] = LoadActor("./Normal.lua", file)
-af[#af+1] = LoadActor("./RainbowMode.lua", file)
-af[#af+1] = LoadActor("./Static.lua", file)
-af[#af+1] = LoadActor("./Technique.lua", file)
+af[#af+1] = Def.Sprite{
+	Texture=THEME:GetPathG("", "VOLT26/SharedBackground.png"),
+	InitCommand=function(self)
+		self:Center():scaletoclipped(_screen.w, _screen.h)
+	end,
+}
+af[#af+1] = LoadActor("./RainbowMode.lua", THEME:GetPathG("", "VOLT26/SharedBackground.png"))
 
 return af